Move resource tests to own file
